refactor: 디렉터리 네이밍 변경
